rename component to counter
