+++ template-component/example/convex/convex.config.ts
@@ -1,5 +1,5 @@
 import { defineApp } from "convex/server";
-import component from "../../src/component/convex.config";
+import component from "../../src/counter/convex.config";
 
 const app = defineApp();
 app.use(component, { name: "counter" });
