+++ template-component/src/counter/_generated/api.d.ts
