+++ template-component/src/counter/_generated/api.js
