+++ template-component/src/counter/_generated/dataModel.d.ts
