+++ template-component/src/counter/_generated/server.d.ts
