+++ template-component/src/counter/_generated/server.js
