+++ template-component/src/counter/convex.config.ts
