+++ template-component/src/counter/public.ts
