+++ template-component/src/counter/schema.ts
